UI: Improve performance of marking all inbox messages as read
Change-type: patch
